refactor: Move pkgs: storage/level/local/writer/writernode -> storage/level/local/writer
